requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.0
seaborn>=0.12.0
plotly>=5.14.0
dash>=2.9.0
tf2onnx>=1.14.0
pyserial>=3.5
